rename etc util to widget util
