+++ src/utils/mod.rs
@@ -1,3 +1,3 @@
 pub mod terminal;
 pub mod align;
-pub mod etc;
+pub mod widget;
